IntervalCoordinateComparator needs to be public
