+++ src/main/java/htsjdk/samtools/util/IntervalList.java
@@ -899,56 +899,3 @@ public class IntervalList implements Iterable<Interval> {
         return result;
     }
 }
-
-/**
- * Comparator that orders intervals based on their sequence index, by coordinate
- * then by strand and finally by name.
- */
-class IntervalCoordinateComparator implements Comparator<Interval>, Serializable {
-    private static final long serialVersionUID = 1L;
-
-    private final SAMFileHeader header;
-
-    /**
-     * Constructs a comparator using the supplied sequence header.
-     */
-    IntervalCoordinateComparator(final SAMFileHeader header) {
-        this.header = header;
-    }
-
-    @Override
-    public int compare(final Interval lhs, final Interval rhs) {
-        final int lhsIndex = this.header.getSequenceIndex(lhs.getContig());
-        final int rhsIndex = this.header.getSequenceIndex(rhs.getContig());
-        int retval = lhsIndex - rhsIndex;
-
-        if (retval == 0) {
-            retval = lhs.getStart() - rhs.getStart();
-        }
-        if (retval == 0) {
-            retval = lhs.getEnd() - rhs.getEnd();
-        }
-        if (retval == 0) {
-            if (lhs.isPositiveStrand() && rhs.isNegativeStrand()) {
-                retval = -1;
-            } else if (lhs.isNegativeStrand() && rhs.isPositiveStrand()) {
-                retval = 1;
-            }
-        }
-        if (retval == 0) {
-            if (lhs.getName() == null) {
-                if (rhs.getName() == null) {
-                    return 0;
-                } else {
-                    return -1;
-                }
-            } else if (rhs.getName() == null) {
-                return 1;
-            } else {
-                return lhs.getName().compareTo(rhs.getName());
-            }
-        }
-
-        return retval;
-    }
-}
